@@ -6,26 +6,16 @@ using Services.Interface.Service.Authorization;
 
 namespace Services.Service.Authorization;
 
-public class AuthorizationService : IAuthorizationService
+public class AuthorizationService(SignInManager<ApplicationUser> signInManager) : IAuthorizationService
 {
-    private readonly SignInManager<ApplicationUser> _signInManager;
-
-    public AuthorizationService(SignInManager<ApplicationUser> signInManager)
-    {
-        _signInManager = signInManager;
-    }
-
     public bool IsSignedIn(ClaimsPrincipal claimsPrincipal)
     {
-        return _signInManager.IsSignedIn(claimsPrincipal);
+        return signInManager.IsSignedIn(claimsPrincipal);
     }
 
     public bool IsLibraryStaff(ClaimsPrincipal claimsPrincipal)
     {
-        if (claimsPrincipal == null)
-        {
-            throw new ArgumentNullException(nameof(claimsPrincipal));
-        }
+        ArgumentNullException.ThrowIfNull(claimsPrincipal);
 
         return claimsPrincipal.IsInRole(UserRoles.Librarian.ToString()) ||
                claimsPrincipal.IsInRole(UserRoles.AssistantLibrarian.ToString()) ||
