@@ -5,19 +5,12 @@ using Services.Interface.Service.Catalog;
 
 namespace Application.Controllers;
 
-public class CatalogController : Controller
+public class CatalogController(ICatalogService catalogService) : Controller
 {
-    private readonly ICatalogService _catalogService;
-
-    public CatalogController(ICatalogService catalogService)
-    {
-        _catalogService = catalogService;
-    }
-
     [AllowAnonymous]
     public async Task<IActionResult> Index()
     {
-        var catalogViewModel = await _catalogService.GetBooksAsync(User);
+        var catalogViewModel = await catalogService.GetBooksAsync(User);
         
         return View(catalogViewModel);
     }
@@ -25,7 +18,7 @@ public class CatalogController : Controller
     [AllowAnonymous]
     public async Task<IActionResult> FilterBooks(BookFilter bookFilter)
     {
-        var catalogViewModel = await _catalogService.GetBooksAsync(bookFilter, User);
+        var catalogViewModel = await catalogService.GetBooksAsync(bookFilter, User);
         return View("Index", catalogViewModel);
     }
 }
