@@ -8,10 +8,6 @@ namespace Data.Data;
 
 public class ApplicationDbContext : IdentityDbContext<ApplicationUser>
 {
-    public ApplicationDbContext()
-    {
-    }
-
     public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options)
         : base(options)
     {
@@ -20,10 +16,10 @@ public class ApplicationDbContext : IdentityDbContext<ApplicationUser>
     // User
     public DbSet<ApplicationUser>? ApplicationUser { get; set; }
     public DbSet<IdentityRole>? Role { get; set; }
-    public DbSet<Book> Book { get; set; }
-    public DbSet<Author> Author { get; set; }
-    public DbSet<Publisher> Publisher { get; set; }
-    public DbSet<Genre> Genre { get; set; }
+    public DbSet<Book>? Book { get; set; }
+    public DbSet<Author>? Author { get; set; }
+    public DbSet<Publisher>? Publisher { get; set; }
+    public DbSet<Genre>? Genre { get; set; }
     
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
